test(web): split app suite by bounded context
